refactor: 코드 분리 및 디자인 개선
